fix: fix bug - 뱃지를 획득할때의 뱃지 규격이 적절하지 않았던 문제를 수정합니다. - 운동 시간을 서버로부터 받을때, 규격이 적절하지 않아 에러가 발생하였던점을 수정합니다. - 운동분석과 운동일지 화면에서 운동을 하나도 완료하지 않았을때에 대한 대응이 적절하지않아 해당 문제를 수정합니다.

+++ core/network/src/main/kotlin/com/gradation/lift/network/dto/history/GetHistoryByHistoryIdResponseDto.kt
@@ -22,9 +22,9 @@ data class GetHistoryByHistoryIdResponseDto(
             historyId = it.value.first().historyId,
             comment = it.value.first().comment,
             score = it.value.first().score,
-            restTime = fromSecondOfDay(it.value.first().restTime),
-            workTime = fromSecondOfDay(it.value.first().workTime),
-            totalTime = fromSecondOfDay(it.value.first().totalTime),
+            restTime = fromSecondOfDay(it.value.first().restTime.toInt()),
+            workTime = fromSecondOfDay(it.value.first().workTime.toInt()),
+            totalTime = fromSecondOfDay(it.value.first().totalTime.toInt()),
             historyTimeStamp = parse(it.value.first().historyTimeStamp),
             historyRoutine = it.value.map { history ->
                 HistoryRoutine(
